+++ src/Other/Strategies/Percentile.php
@@ -7,22 +7,43 @@ use InvalidArgumentException;
 use RuntimeException;
 
 /**
- * Mean
+ * Percentile
  *
- * This strategy always predicts the mean of the fitted data.
+ * A strategy that always guesses the p-th percentile of the fitted data.
  *
  * @category    Machine Learning
  * @package     Rubix/ML
  * @author      Andrew DalPino
  */
-class Mean implements Continuous
+class Percentile implements Continuous
 {
     /**
-     * The mean of the fitted values.
+     * The percentile of the fitted data to use as a guess.
+     *
+     * @var float
+     */
+    protected $p;
+
+    /**
+     * The pth percentile of the fitted data.
      *
      * @var float|null
      */
-    protected $mean;
+    protected $percentile;
+
+    /**
+     * @param float $p
+     * @throws \InvalidArgumentException
+     */
+    public function __construct(float $p = 50.0)
+    {
+        if ($p < 0. or $p > 100.) {
+            throw new InvalidArgumentException('Percentile must be between'
+                . " 0 and 100, $p given.");
+        }
+
+        $this->p = $p;
+    }
 
     /**
      * Fit the guessing strategy to a set of values.
@@ -37,21 +58,21 @@ class Mean implements Continuous
                 . ' to at least 1 value.');
         }
 
-        $this->mean = Stats::mean($values);
+        $this->percentile = Stats::percentile($values, $this->p);
     }
 
     /**
-     * Make a continuous guess.
+     * Make a guess.
      *
      * @throws \RuntimeException
      * @return float
      */
     public function guess() : float
     {
-        if ($this->mean === null) {
+        if ($this->percentile === null) {
             throw new RuntimeException('Strategy has not been fitted.');
         }
 
-        return $this->mean;
+        return $this->percentile;
     }
 }