+++ tests/Other/Strategies/MeanTest.php
@@ -27,7 +27,7 @@ class MeanTest extends TestCase
         $this->assertInstanceOf(Strategy::class, $this->strategy);
     }
 
-    public function test_make_guess()
+    public function test_guess()
     {
         $this->strategy->fit($this->values);