+++ tests/Other/Strategies/PercentileTest.php
@@ -4,10 +4,10 @@ namespace Rubix\ML\Tests\Other\Strategies;
 
 use Rubix\ML\Other\Strategies\Strategy;
 use Rubix\ML\Other\Strategies\Continuous;
-use Rubix\ML\Other\Strategies\BlurryPercentile;
+use Rubix\ML\Other\Strategies\Percentile;
 use PHPUnit\Framework\TestCase;
 
-class BlurryPercentileTest extends TestCase
+class PercentileTest extends TestCase
 {
     protected $values;
 
@@ -17,17 +17,17 @@ class BlurryPercentileTest extends TestCase
     {
         $this->values = [1, 2, 3, 4, 5];
 
-        $this->strategy = new BlurryPercentile(50., 0.3);
+        $this->strategy = new Percentile(50.);
     }
 
     public function test_build_strategy()
     {
-        $this->assertInstanceOf(BlurryPercentile::class, $this->strategy);
+        $this->assertInstanceOf(Percentile::class, $this->strategy);
         $this->assertInstanceOf(Continuous::class, $this->strategy);
         $this->assertInstanceOf(Strategy::class, $this->strategy);
     }
 
-    public function test_make_guess()
+    public function test_guess()
     {
         $this->strategy->fit($this->values);
 
@@ -36,8 +36,8 @@ class BlurryPercentileTest extends TestCase
         $this->assertThat(
             $guess,
             $this->logicalAnd(
-                $this->greaterThanOrEqual(0),
-                $this->lessThanOrEqual(6)
+                $this->greaterThanOrEqual(1),
+                $this->lessThanOrEqual(5)
             )
         );
     }